Rough skeleton of the generalized pythagoras tree and accompanying data structure
Rough skeleton of the generalized pythagoras tree. The core of the algorithm and its form are there but some functionality should be split in to seperate functions and some functions lack their required logic still.

There is also a rough data structure for storing a tree. It's currently inefficient and not fully featured. However, it was required to have one. Can be adapted or exchanged for one that is more suited to the rest of the project.